+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.0.15
 - Add user-configurable OpenCore DEBUG builds
 - Add user-configurable Wifi and GPU patches
+- Fix ThirdPartyDrives model detection
 
 ## 0.0.14
 - Enable ThirdPartyDrives to aid with hibernation on 3rd party SATA drives